Reorg species to be consistent with other types
