+++ charsheets/species/orc.py
@@ -1,11 +1,11 @@
 from charsheets.abilities.base_ability import BaseAbility
 from charsheets.constants import Ability
-from charsheets.species import Species
+from charsheets.species.base_species import BaseSpecies
 from charsheets.abilities import Darkvision120
 
 
 #############################################################################
-class Orc(Species):
+class Orc(BaseSpecies):
     #########################################################################
     def species_abilities(self) -> set[BaseAbility]:
         results: set[BaseAbility] = {RelentlessEndurance(), Darkvision120(), AdrenalinRush()}
