+++ examples/rugor.py
@@ -2,7 +2,7 @@ from charsheets.abilities import AbilityScoreImprovement
 from charsheets.classes import PathOfTheBeserker
 from charsheets.constants import Skill, Armour, Stat, Feat, Tool
 from charsheets.origins import Artisan
-from charsheets.species.dwarf import Dwarf
+from charsheets.species import Dwarf
 from charsheets.weapons import Shortbow, Warhammer
 from charsheets.armour import Scale
